refactor: rename gitTreeLister module to treeIndexes
