+++ __tests__/integration/adapter/treeIndexScopeRegression.test.ts
@@ -4,8 +4,8 @@ import { rm } from 'node:fs/promises'
 import { afterAll, afterEach, beforeAll, describe, expect, it } from 'vitest'
 
 import GitAdapter from '../../../src/adapter/GitAdapter'
-import { createTreeIndexes } from '../../../src/adapter/gitTreeLister'
 import type { TreeIndex } from '../../../src/adapter/treeIndex'
+import { createTreeIndexes } from '../../../src/adapter/treeIndexes'
 import { MetadataRepository } from '../../../src/metadata/MetadataRepository'
 import { getDefinition } from '../../../src/metadata/metadataManager'
 import type { Config } from '../../../src/types/config'
