+++ __tests__/unit/lib/adapter/ioExecutor.test.ts
@@ -4,8 +4,8 @@ import { PassThrough, Readable, Writable } from 'node:stream'
 import type { Ignore } from 'ignore'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { EscalateToStreamingSignal } from '../../../../src/adapter/gitBlobReader'
-import type { TreeIndexes } from '../../../../src/adapter/gitTreeLister'
 import IOExecutor from '../../../../src/adapter/ioExecutor'
+import type { TreeIndexes } from '../../../../src/adapter/treeIndexes'
 import type { CopyOperation } from '../../../../src/types/handlerResult'
 import { CopyOperationKind } from '../../../../src/types/handlerResult'
 import { outputFile } from '../../../../src/utils/fsUtils'
