+++ src/post-processor/postProcessorManager.ts
@@ -1,5 +1,5 @@
 'use strict'
-import type { TreeIndexes } from '../adapter/gitTreeLister.js'
+import type { TreeIndexes } from '../adapter/treeIndexes.js'
 import { MetadataRepository } from '../metadata/MetadataRepository.js'
 import type { Config } from '../types/config.js'
 import type { HandlerResult } from '../types/handlerResult.js'
