+++ src/service/containedDecomposedHandler.ts
@@ -1,6 +1,6 @@
 'use strict'
 import { join, ParsedPath, parse } from 'node:path/posix'
-import type { TreeIndexes } from '../adapter/gitTreeLister.js'
+import type { TreeIndexes } from '../adapter/treeIndexes.js'
 import { PATH_SEP } from '../constant/fsConstants.js'
 import { METAFILE_SUFFIX } from '../constant/metadataConstants.js'
 import type { Config } from '../types/config.js'
