+++ src/service/sharedFolderHandler.ts
@@ -1,6 +1,6 @@
 'use strict'
 import { join, parse } from 'node:path/posix'
-import type { TreeIndexes } from '../adapter/gitTreeLister.js'
+import type { TreeIndexes } from '../adapter/treeIndexes.js'
 import { METAFILE_SUFFIX } from '../constant/metadataConstants.js'
 import type { Config } from '../types/config.js'
 import type { HandlerResult } from '../types/handlerResult.js'
