+++ src/utils/metadataBoundaryResolver.ts
@@ -1,7 +1,7 @@
 'use strict'
 import { dirname, parse } from 'node:path/posix'
 
-import type { TreeIndexes } from '../adapter/gitTreeLister.js'
+import type { TreeIndexes } from '../adapter/treeIndexes.js'
 import { PATH_SEP } from '../constant/fsConstants.js'
 import { METAFILE_SUFFIX } from '../constant/metadataConstants.js'
 import { MetadataRepository } from '../metadata/MetadataRepository.js'
